+++ src/Instagram.php
@@ -1786,13 +1786,12 @@ class Instagram implements ExperimentsInterface
             if ($this->loginAttemptCount === 0 && !self::$skipLoginFlowAtMyOwnRisk) {
                 $this->_sendPreLoginFlow();
             }
-            /* THIS IS NOT USED ANYMORE IN BLOKS LOGIN
-            else {
+            // THIS IS NOT USED ANYMORE IN BLOKS LOGIN
+            if (self::$useBloksLogin === false) {
                 $mobileConfigResponse = $this->internal->getMobileConfig(true)->getHttpResponse();
                 $this->settings->set('public_key', $mobileConfigResponse->getHeaderLine('ig-set-password-encryption-pub-key'));
                 $this->settings->set('public_key_id', $mobileConfigResponse->getHeaderLine('ig-set-password-encryption-key-id'));
             }
-            */
 
             $this->event->sendStringImpressions(['2131231876' => 1, '2131231882' => 1, '2131886885' => 2, '2131887195' => 1, '2131887196' => 1, '2131888193' => 4, '2131888472' => 1, '2131890367' => 1, '2131891325' => 1, '2131892179' => 1, '2131892669' => 1, '2131892673' => 1, '2131893765' => 1, '2131893766' => 1, '2131893767' => 1, '2131893768' => 1, '2131893769' => 1, '2131893770' => 1, '2131893771' => 1, '2131893772' => 1, '2131893773' => 1, '2131893774' => 1, '2131893775' => 1, '2131893776' => 1, '2131893777' => 1, '2131893778' => 1, '2131893779' => 1, '2131893780' => 1, '2131893781' => 1, '2131893782' => 1, '2131893783' => 1, '2131893784' => 1, '2131893785' => 1, '2131893788' => 1, '2131893789' => 1, '2131893790' => 1, '2131893791' => 2, '2131893792' => 1, '2131893793' => 1, '2131893806' => 1, '2131893898' => 1, '2131894010' => 1, '2131894018' => 1, '2131896911' => 1, '2131898165' => 1]);
             $this->event->sendFlowSteps('login', 'log_in_username_focus', $waterfallId, $startTime);
@@ -2937,11 +2936,12 @@ class Instagram implements ExperimentsInterface
                 /*
                 //THIS WAS USED IN PRELOGIN FOR OBTAINING DEVICE EXPERIMENTS AND PUBLIC KEY TO ENCRYPT PASSWORDS
                 //SEEMS IT IS NOT BEING USED ANYMORE WITH BLOKS LOGIN
-
-                $mobileConfigResponse = $this->internal->getMobileConfig(true)->getHttpResponse();
-                $this->settings->set('public_key', $mobileConfigResponse->getHeaderLine('ig-set-password-encryption-pub-key'));
-                $this->settings->set('public_key_id', $mobileConfigResponse->getHeaderLine('ig-set-password-encryption-key-id'));
                 */
+                if (self::$useBloksLogin === false) {
+                    $mobileConfigResponse = $this->internal->getMobileConfig(true)->getHttpResponse();
+                    $this->settings->set('public_key', $mobileConfigResponse->getHeaderLine('ig-set-password-encryption-pub-key'));
+                    $this->settings->set('public_key_id', $mobileConfigResponse->getHeaderLine('ig-set-password-encryption-key-id'));
+                }
 
                 //$this->internal->bootstrapMsisdnHeader();
                 /*
@@ -2987,7 +2987,9 @@ class Instagram implements ExperimentsInterface
                 $this->account->getNamePrefill();
             }
             // WAS USED BEFORE BLOKS LOGIN
-            //$this->internal->getMobileConfig(true);
+            if (self::$useBloksLogin === false) {
+                $this->internal->getMobileConfig(true);
+            }
 
             /* QE SYNC DISABLED
             try {